refactor(#214): split world.cpp into per-concern files
