+++ clients/silencer/src/world/world.cpp
@@ -105,16 +105,6 @@ World::~World(){
 	}
 }
 
-void World::DoNetwork(void){
-	if(mode == AUTHORITY){
-		DoNetwork_Authority();
-	}else{
-		DoNetwork_Replica();
-	}
-	lobby.DoNetwork();
-	lagsimulator.Process(*this);
-}
-
 void World::Tick(void){
 	if(mode == AUTHORITY){
 		SendSnapshots();
@@ -242,2688 +232,7 @@ void World::Tick(void){
 	}
 }
 
-void World::TickObjects(void){
-	for(std::list<Object *>::reverse_iterator i = objectlist.rbegin(); i != objectlist.rend(); i++){
-		Object * object = (*i);
-		bool peercontrolled = false;
-		bool localpeercontrolled = false;
-		if(object->iscontrollable){
-			for(int i = 0; i < maxpeers; i++){
-				Peer * peer = peerlist[i];
-				if(peer && !peer->isbot){
-					for(std::list<Uint16>::iterator it = peer->controlledlist.begin(); it != peer->controlledlist.end(); it++){
-						if((*it) == object->id){
-							peercontrolled = true;
-							if(peer == peerlist[localpeerid]){
-								localpeercontrolled = true;
-							}
-						}
-					}
-				}
-			}
-		}
-		if(!(object->requiresmaptobeloaded && !map.loaded) && !object->wasdestroyed){
-			if((!IsAuthority() && !localpeercontrolled) || (IsAuthority() && !peercontrolled)){
-				object->oldx = object->x;
-				object->oldy = object->y;
-				object->Tick(*this);
-				object->lasttick = tickcount;
-			}/*else
-			if(!IsAuthority() && !localpeercontrolled){
-				// Tick everything to do the sounds and effects, then go back
-				Serializer olddata;
-				object->Serialize(Serializer::WRITE, olddata);
-				//LoadSnapshot(*data, false);
-				//replaying = true;
-				object->oldx = object->x;
-				object->oldy = object->y;
-				object->Tick(*this);
-				object->lasttick = tickcount;
-				//replaying = false;
-				object->Serialize(Serializer::READ, olddata);
-				//
-			}*/
-		}
-	}
-	DestroyMarkedObjects();
-	Player * localplayer = GetPeerPlayer(localpeerid);
-	if(localplayer){
-		audio.UpdateAllVolumes(*this, localplayer->x, localplayer->y, GASLoader::Get().world.audioRange);
-	}else{
-		audio.UpdateAllVolumes(*this, replay.x, replay.y, GASLoader::Get().world.audioRange);
-	}
-}
-
 void World::SetVersion(const char * version){
 	strcpy(World::version, version);
 }
 
-void World::DoNetwork_Authority(void){
-	Serializer data(10000);
-	sockaddr_in senderaddr;
-	socklen_t senderaddrsize = sizeof(senderaddr);
-	int received;
-	while((received = recvfrom(sockethandle, data.data, data.size, 0, (sockaddr *)&senderaddr, &senderaddrsize)) > 0){
-		data.offset = received * 8;
-		data.readoffset = 0;
-		totalbytesread += received;
-		Peer * peer = FindPeer(senderaddr);
-		if(peer){
-			peer->lastpacket = SDL_GetTicks();
-		}
-		char code;
-		data.Get(code);
-		switch(code){
-			case MSG_CONNECT:{
-				//printf("MSG_CONNECT received\n");
-				Serializer response;
-				Uint8 code = MSG_CONNECT;
-				response.Put(code);
-				if(gameplaystate == INLOBBY || gameplaystate == INGAME){
-					char * host = inet_ntoa(senderaddr.sin_addr);
-					unsigned short port = ntohs(senderaddr.sin_port);
-					Uint8 agency;
-					data.Get(agency);
-					Uint32 accountid;
-					data.Get(accountid);
-					Uint8 passwordsize;
-					data.Get(passwordsize);
-					char temp[256];
-					memset(temp, 0, sizeof(temp));
-					for(int i = 0; i < passwordsize; i++){
-						data.Get(temp[i]);
-					}
-					bool observerRequest = data.GetBit();
-					bool canjoin = true;
-					if(strcmp(gameinfo.password, temp) != 0){
-						if(!(dedicatedserver.active && accountid == dedicatedserver.accountid)){
-							//printf("bad password\n");
-							canjoin = false;
-						}
-					}
-					// Find a parked peer to rebind, if any. Done before the
-					// maxplayers gate so a rejoiner reclaiming their existing
-					// slot isn't rejected for a full lobby.
-					Peer * rejoinpeer = 0;
-					if(accountid != 0){
-						for(unsigned int i = 1; i < maxpeers; i++){
-							if(peerlist[i] && peerlist[i]->disconnected && peerlist[i]->accountid == accountid){
-								rejoinpeer = peerlist[i];
-								break;
-							}
-						}
-					}
-					if(dedicatedserver.active){
-						if(dedicatedserver.IsBanned(accountid)){
-							canjoin = false;
-						}
-						if(!rejoinpeer && !observerRequest && peercount >= gameinfo.maxplayers){
-							canjoin = false;
-						}
-					}
-					if(canjoin && observerRequest && !gameinfo.spectatable){
-						// defense-in-depth: button gating already prevents this in normal flow
-						canjoin = false;
-					}
-					if(canjoin && gameplaystate == INGAME && !rejoinpeer && !observerRequest){
-						// mid-game connects are only for rejoiners or observers
-						canjoin = false;
-					}
-					if(canjoin && rejoinpeer){
-						rejoinpeer->ip = ntohl(inet_addr(host));
-						rejoinpeer->port = port;
-						rejoinpeer->lastpacket = SDL_GetTicks();
-						rejoinpeer->disconnected = false;
-						for(std::list<Uint16>::iterator it = rejoinpeer->controlledlist.begin(); it != rejoinpeer->controlledlist.end(); it++){
-							Object * obj = GetObjectFromId(*it);
-							if(obj && obj->type == ObjectTypes::PLAYER){
-								Player * p = static_cast<Player *>(obj);
-								map.RandomPlayerStartLocation(*this, p->x, p->y);
-								p->oldx = p->x;
-								p->oldy = p->y;
-								p->health = p->maxhealth;
-								p->shield = p->maxshield;
-								p->state = Player::DEPLOYING;
-								p->state_i = GASLoader::Get().player.deployWaitTicks;
-								p->draw = false;
-								p->collidable = false;
-							}
-						}
-						response.PutBit(true);
-						response.Put(rejoinpeer->id);
-						SendGameInfo(rejoinpeer->id);
-						SendPeerList();
-					}else if(canjoin && observerRequest){
-						Peer * newpeer = AddPeer(host, port, agency, accountid, true);
-						if(newpeer){
-							newpeer->observer = true;
-							response.PutBit(true);
-							response.Put(newpeer->id);
-							SendGameInfo(newpeer->id);
-							SendPeerList();
-						}else{
-							response.PutBit(false);
-						}
-					}else if(canjoin){
-						Peer * newpeer = AddPeer(host, port, agency, accountid);
-						if(newpeer){
-							if(dedicatedserver.active){
-								lobby.GetUserInfo(newpeer->accountid);
-								if(newpeer->accountid == dedicatedserver.accountid){
-									//printf("host connected\n");
-									newpeer->ishost = true;
-									newpeer->gameinfoloaded = true;
-									newpeer->mapdownloaded = true;
-								}
-							}
-							response.PutBit(true);
-							response.Put(newpeer->id);
-							if(!newpeer->ishost){
-								SendGameInfo(newpeer->id);
-							}
-							if(replay.IsRecording()){
-								replay.WriteNewPeer(agency, accountid);
-							}
-						}else{
-							//printf("couldnt add peer\n");
-							response.PutBit(false);
-						}
-					}else{
-						response.PutBit(false);
-					}
-				}else{
-					response.PutBit(false);
-				}
-				Peer temppeer;
-				temppeer.ip = ntohl(senderaddr.sin_addr.s_addr);
-				temppeer.port = ntohs(senderaddr.sin_port);
-				SendPacket(&temppeer, response.data, response.BitsToBytes(response.offset));
-			}break;
-			case MSG_INPUT:{ // client sending input
-				if(peer && !peer->observer && gameplaystate == INGAME){
-					totalinputpackets++;
-					peer->totalinputs++;
-					Serializer * inputcopy = new Serializer;
-					inputcopy->Copy(data);
-					inputqueue[peer->id].push_back(inputcopy);
-					if(!peer->firstinputtime){
-						peer->firstinputtime = tickcount;
-					}
-					if(replay.IsRecording()){
-						replay.WriteInputCommand(*this, peer->id, data);
-					}
-				}
-			}break;
-			case MSG_PEERLIST:{ // peerlist requested
-				if(peer){
-					SendPeerList(peer->id);
-				}
-			}break;
-			case MSG_DISCONNECT:{ // disconnect
-				if(peer){
-					HandleDisconnect(peer->id);
-				}
-			}break;
-			case MSG_PING:{ // ping
-				//printf("received ping from %s:%d\n", inet_ntoa(senderaddr.sin_addr), ntohs(senderaddr.sin_port));
-				if(peer){
-					Uint32 pingid;
-					data.Get(pingid);
-					Serializer response;
-					Uint8 code = MSG_PONG;
-					response.Put(code);
-					response.Put(pingid);
-					SendPacket(peer, response.data, response.BitsToBytes(response.offset));
-				}
-			}break;
-			case MSG_PONG:{ // pong
-
-			}break;
-			case MSG_GAMEINFO:{
-				//printf("Received MSG_GAMEINFO\n");
-				if(peer){
-					if(peer->ishost){
-						//printf("loading game info from host\n");
-						gameinfo.Serialize(Serializer::READ, data);
-						if(replay.IsRecording()){
-							replay.WriteGameInfo(gameinfo);
-						}
-						Peer * localpeer = peerlist[localpeerid];
-						if(localpeer){
-							localpeer->gameinfoloaded = true;
-						}
-						for(int i = 0; i < maxpeers; i++){
-							Peer * peer = peerlist[i];
-							if(peer && peer->id != localpeerid && !peer->ishost){
-								SendGameInfo(peer->id);
-							}
-						}
-					}else{
-						peer->gameinfoloaded = true;
-						//SendGameInfo(peer->id);
-					}
-				}
-			}break;
-			case MSG_READY:{
-				if(peer){
-					if(peer->isready){
-						peer->isready = false;
-					}else{
-						peer->isready = true;
-					}
-					SendPeerList();
-					if(!peer->gameinfoloaded){
-						SendGameInfo(peer->id);
-					}
-				}
-			}break;
-			case MSG_CHAT:{
-				if(peer){
-					Player * player = GetPeerPlayer(peer->id);
-					Uint8 to;
-					data.Get(to);
-					if(peer->observer && to == 1){
-						to = 0;
-					}
-					Serializer response;
-					Uint8 code = MSG_CHAT;
-					response.Put(code);
-					response.Put(peer->accountid);
-					data.data[256] = 0;
-					char * msg = &data.data[data.BitsToBytes(data.readoffset)];
-					for(int i = 0; i < strlen(msg); i++){
-						response.Put(msg[i]);
-					}
-					msg[strlen(msg)] = 0;
-					if(replay.IsRecording()){
-						replay.WriteChat(peer->id, to, msg);
-					}
-					char nullend = 0;
-					response.Put(nullend);
-					if(to == 1){ // send to team
-						if(player){
-							Team * team = player->GetTeam(*this);
-							if(team){
-								for(int i = 0; i < team->numpeers; i++){
-									if(peerlist[team->peers[i]] && team->peers[i] != localpeerid){
-										SendPacket(peerlist[team->peers[i]], response.data, response.BitsToBytes(response.offset));
-									}
-								}
-							}
-						}
-					}else{
-						for(int i = 0; i < maxpeers; i++){
-							if(peerlist[i] && i != localpeerid){
-								SendPacket(peerlist[i], response.data, response.BitsToBytes(response.offset));
-							}
-						}
-					}
-				}
-			}break;
-			case MSG_STATION:{
-				if(peer){
-					Uint8 subcode;
-					data.Get(subcode);
-					Uint8 id;
-					data.Get(id);
-					switch(subcode){
-						case STA_BUY:{
-							data.Get(id);
-							Player * player = GetPeerPlayer(peer->id);
-							if(player){
-								player->BuyItem(*this, id);
-							}
-							if(replay.IsRecording()){
-								replay.WriteStation(peer->id, Replay::STA_BUY, id);
-							}
-						}break;
-						case STA_REPAIR:{
-							Player * player = GetPeerPlayer(peer->id);
-							if(player){
-								player->RepairItem(*this, id);
-							}
-							if(replay.IsRecording()){
-								replay.WriteStation(peer->id, Replay::STA_REPAIR, id);
-							}
-						}break;
-						case STA_VIRUS:{
-							Player * player = GetPeerPlayer(peer->id);
-							if(player){
-								player->VirusItem(*this, id);
-							}
-							if(replay.IsRecording()){
-								replay.WriteStation(peer->id, Replay::STA_VIRUS, id);
-							}
-						}break;
-					}
-					/*Uint8 id;
-					data.Get(id);
-					Player * player = GetPeerPlayer(peer->id);
-					if(player){
-						player->BuyItem(*this, id);
-					}
-					if(replay.IsRecording()){
-						replay.WriteStation(peer->id, Replay::STA_BUY, id);
-					}*/
-				}
-			}break;
-			/*case MSG_REPAIR:{
-				if(peer){
-					Uint8 id;
-					data.Get(id);
-					Player * player = GetPeerPlayer(peer->id);
-					if(player){
-						player->RepairItem(*this, id);
-					}
-					if(replay.IsRecording()){
-						replay.WriteStation(peer->id, Replay::STA_REPAIR, id);
-					}
-				}
-			}break;
-			case MSG_VIRUS:{
-				if(peer){
-					Uint8 id;
-					data.Get(id);
-					Player * player = GetPeerPlayer(peer->id);
-					if(player){
-						player->VirusItem(*this, id);
-					}
-					if(replay.IsRecording()){
-						replay.WriteStation(peer->id, Replay::STA_VIRUS, id);
-					}
-				}
-			}break;*/
-			case MSG_CHANGETEAM:{
-				if(peer && gameplaystate == INLOBBY){
-					ChangeTeam(peer->id);
-					if(replay.IsRecording()){
-						replay.WriteChangeTeam(peer->id);
-					}
-				}
-			}break;
-			case MSG_SETAGENCY:{
-				if(peer && gameplaystate == INLOBBY){
-					Uint8 newagency;
-					data.Get(newagency);
-					Team * peerteam = GetPeerTeam(peer->id);
-					if(peerteam && peerteam->agency != newagency){
-						peerteam->RemovePeer(peer->id);
-						peer->isready = false;
-						FindTeamForPeer(*peer, newagency);
-						SendPeerList();
-					}
-				}
-			}break;
-			case MSG_TECH:{
-				if(peer && gameplaystate == INLOBBY){
-					Uint32 techchoices;
-					data.Get(techchoices);
-					SetTech(peer->id, techchoices);
-					if(replay.IsRecording()){
-						replay.WriteSetTech(peer->id, techchoices);
-					}
-				}
-			}break;
-			case MSG_EXISTS:{
-				if(peer){
-					Serializer response;
-					Uint8 code = MSG_REMOVE;
-					response.Put(code);
-					while(data.MoreBytesToRead()){
-						Uint16 objectid;
-						data.Get(objectid);
-						if(!GetObjectFromId(objectid)){
-							response.Put(objectid);
-						}
-					}
-					SendPacket(peer, response.data, response.BitsToBytes(response.offset));
-				}
-			}break;
-			case MSG_MAP:{
-				if(peer){
-					Uint8 subcode;
-					data.Get(subcode);
-					switch(subcode){
-						case MAP_DOWNLOADED:{
-							peer->mapdownloaded = true;
-							SendPeerList();
-							BroadcastTriggerState();
-						}break;
-						case MAP_GETCHUNK:{
-							Uint32 offset;
-							data.Get(offset);
-							PutMapChunk(offset, *peer);
-						}break;
-						case MAP_PUTCHUNK:{
-							if(peer->ishost){
-								Uint32 offset;
-								data.Get(offset);
-								Uint32 size;
-								data.Get(size);
-								StoreMapChunk((unsigned char *)&data.data[data.BitsToBytes(data.readoffset)], offset, size);
-							}
-						}break;
-					}
-				}
-			}break;
-			case MSG_KICK:{
-				if(dedicatedserver.active && senderaddr.sin_addr.s_addr == inet_addr(dedicatedserver.lobbyaddress)){
-					Uint32 accountid;
-					data.Get(accountid);
-					for(int i = 0; i < maxpeers; i++){
-						Peer * p = peerlist[i];
-						if(p && p->accountid == accountid){
-							if(gameplaystate == World::INGAME){
-								KillByGovt(*p);
-							}
-							HandleDisconnect(p->id, true);
-							break;
-						}
-					}
-				}
-			}break;
-		}
-	}
-	if(!replay.IsPlaying()){
-		Uint32 tickcheck = SDL_GetTicks();
-		for(int i = 0; i < maxpeers; i++){
-			if(peerlist[i]){
-				if(i != localpeerid && !peerlist[i]->isbot && !peerlist[i]->disconnected && peerlist[i]->lastpacket < tickcheck && tickcheck - peerlist[i]->lastpacket >= peertimeout){
-					HandleDisconnect(i);
-				}
-			}
-		}
-	}
-}
-
-void World::DoNetwork_Replica(void){
-	if(!peerlist[authoritypeer]){
-		return;
-	}
-	if(SDL_GetTicks() - lastpingsent >= (Uint32)GASLoader::Get().gameengine.pingIntervalMs){
-		SendPing();
-	}
-	Serializer data(10000); // hopefully snapshots dont get larger than this
-	sockaddr_in senderaddr;
-	socklen_t senderaddrsize = sizeof(senderaddr);
-	Peer * peer = 0;
-	int received;
-	while((received = recvfrom(sockethandle, data.data, data.size, 0, (sockaddr *)&senderaddr, &senderaddrsize)) > 0){
-		//printf("received data from %s:%d\n", inet_ntoa(senderaddr.sin_addr), ntohs(senderaddr.sin_port));
-		if(peerlist[authoritypeer]->ip == ntohl(senderaddr.sin_addr.s_addr) && peerlist[authoritypeer]->port == ntohs(senderaddr.sin_port)){
-			peer = peerlist[authoritypeer];
-			peer->lastpacket = SDL_GetTicks();
-		}
-		data.offset = received * 8;
-		data.readoffset = 0;
-		totalbytesread += received;
-		char code;
-		data.Get(code);
-		switch(code){
-			case MSG_CONNECT:{ // connect response
-				//printf("MSG_CONNECT response received from %s:%d\n", inet_ntoa(senderaddr.sin_addr), ntohs(senderaddr.sin_port));
-				if(peer){
-					if(data.GetBit()){
-						data.Get(localpeerid);
-						viewedpeerid = localpeerid;
-						//printf("we are connected, our peer id is %d\n", localpeerid);
-						RequestPeerList();
-					}else{
-						//printf("failed to connect to game\n");
-						// host not accepting connection, password is wrong, or teams are full, etc
-						Disconnect();
-					}
-				}
-			}break;
-			case MSG_SNAPSHOT:{ // snapshot data
-				if(peer){
-					if(state == CONNECTED){
-						totalsnapshots++;
-						Serializer * snapshotcopy = new Serializer;
-						snapshotcopy->Copy(data);
-						snapshotqueue.push_back(snapshotcopy);
-					}
-				}
-			}break;
-			case MSG_PEERLIST:{ // peerlist update
-				if(peer){
-					ReadPeerList(data);
-					// The MSG_CONNECT response (3 bytes) is often dropped by
-					// carrier-grade NAT. Fall back to self-identifying by
-					// accountid: our peer in the authority's peerlist has
-					// accountid == lobby.accountid and id != authoritypeer.
-					if(localpeerid == authoritypeer){
-						for(unsigned int i = 0; i < maxpeers; i++){
-							if(i == authoritypeer) continue;
-							if(peerlist[i] && peerlist[i]->accountid == lobby.accountid){
-								localpeerid = i;
-								viewedpeerid = i;
-								break;
-							}
-						}
-					}
-					Peer * localpeer = peerlist[localpeerid];
-					if(localpeer && localpeer->ishost && !GetAuthorityPeer()->gameinfoloaded && gameinfo.loaded){
-						SendGameInfo(GetAuthorityPeer()->id);
-						//printf("We are host, sending game info\n");
-					}
-					if(state == CONNECTING && localpeer){
-						state = CONNECTED;
-					}
-				}
-			}break;
-			case MSG_DISCONNECT:{ // disconnect
-				if(peer){
-					HandleDisconnect(authoritypeer);
-				}
-			}break;
-			case MSG_PING:{ // ping
-				//printf("received ping from %s:%d\n", inet_ntoa(senderaddr.sin_addr), ntohs(senderaddr.sin_port));
-			}break;
-			case MSG_PONG:{ // pong
-				//printf("received MSG_PONG\n");
-				Uint32 pingid;
-				data.Get(pingid);
-				if(pingid == lastpingid){
-					pingtime = SDL_GetTicks() - lastpingsent;
-					pinghistory[lastpingid % (sizeof(pinghistory) / sizeof(int))] = pingtime;
-				}
-			}break;
-			case MSG_GAMEINFO:{
-				if(peer){
-					//printf("Received MSG_GAMEINFO\n");
-					gameinfo.Serialize(Serializer::READ, data);
-					//printf("password %s\n", gameinfo.password);
-					SendGameInfoLoaded();
-					/*Serializer response;
-					Uint8 code = MSG_GAMEINFO;
-					response.Put(code);
-					SendPacket(GetAuthorityPeer(), response.data, response.BitsToBytes(response.offset));*/
-				}
-			}break;
-			case MSG_CHAT:{
-				Uint32 accountid;
-				data.Get(accountid);
-				DisplayChatMessage(accountid, &data.data[1 + 4]);
-				/*std::string chatmsg(lobby.GetUserInfo(accountid)->name);
-				chatmsg.append(":\xA0");
-				chatmsg.append(&data.data[1 + 4]);
-				
-				char * wrapped = silencer::ui::WordWrapText(chatmsg.c_str(), 36);
-				char * line = strtok(wrapped, "\n");
-				while(line){
-					chatlines.push_back(line);
-					line = strtok(NULL, "\n");
-				}
-				delete[] wrapped;
-				
-				showchat_i = GASLoader::Get().gameengine.chatDisplayTicks;
-				while(chatlines.size() > 5){
-					chatlines.pop_front();
-				}*/
-			}break;
-			case MSG_STATUS:{
-				int size = data.BitsToBytes(data.offset - data.readoffset);
-				char * newstatus = new char[size];
-				memcpy(newstatus, &data.data[data.BitsToBytes(data.readoffset)], size);
-				PushStatusString(newstatus);
-			}break;
-			case MSG_MESSAGE:{
-				char message[1024];
-				strcpy(message, &data.data[data.BitsToBytes(data.readoffset)]);
-				Uint8 time = data.data[data.BitsToBytes(data.readoffset) + strlen(message) + 1];
-				Uint8 type = data.data[data.BitsToBytes(data.readoffset) + strlen(message) + 1 + 1];
-				//printf("GOT MSG_MESSAGE %s %d %d\n", message, time, type);
-				ShowMessage(message, time, type);
-			}break;
-			case MSG_STATS:{
-				if(peer){
-					Peer * localpeer = peerlist[localpeerid];
-					if(localpeer){
-						localpeer->stats.Serialize(Serializer::READ, data);
-						//printf("MSG_STATS received\n");
-						Team * team = GetPeerTeam(localpeer->id);
-						if(team){
-							User * user = lobby.GetUserInfo(localpeer->accountid);
-							char namecopy[64];
-							strcpy(namecopy, user->name);
-							lobby.ForgetUserInfo(localpeer->accountid);
-							user = lobby.GetUserInfo(localpeer->accountid);
-							strcpy(user->name, namecopy);
-							user->statscopy = localpeer->stats;
-							user->statsagency = team->agency;
-							user->teamnumber = team->number;
-						}
-
-					}
-				}
-			}break;
-			case MSG_GOVTKILL:{
-				if(peer){
-					Uint8 peerid = data.data[data.BitsToBytes(data.readoffset)];
-					Player * player = GetPeerPlayer(peerid);
-					if(player){
-						player->KillByGovt(*this);
-					}
-				}
-			}break;
-			case MSG_SOUND:{
-				if(peer){
-					Uint8 volume = data.data[data.BitsToBytes(data.readoffset)];
-					char * name = &data.data[data.BitsToBytes(data.readoffset) + 1];
-					Audio::GetInstance().Play(resources.soundbank[name], volume);
-					//printf("MSG_SOUND %d %s\n", volume, name);
-				}
-			}break;
-			case MSG_REMOVE:{
-				if(peer){
-					while(data.MoreBytesToRead()){
-						Uint16 objectid;
-						data.Get(objectid);
-						Object * object = GetObjectFromId(objectid);
-						if(object){
-							MarkDestroyObject(objectid);
-							//printf("DELETED %d\n", objectid);
-						}
-					}
-				}
-			}break;
-			case MSG_MAP:{
-				//printf("MSG_MAP received\n");
-				if(peer){
-					Uint8 subcode;
-					data.Get(subcode);
-					//printf("%d\n", subcode);
-					switch(subcode){
-						case MAP_GETCHUNK:{
-							Uint32 offset;
-							data.Get(offset);
-							PutMapChunk(offset, *peer);
-						}break;
-						case MAP_PUTCHUNK:{
-							Uint32 offset;
-							data.Get(offset);
-							Uint32 size;
-							data.Get(size);
-							StoreMapChunk((unsigned char *)&data.data[data.BitsToBytes(data.readoffset)], offset, size);
-						}break;
-					}
-				}
-			}break;
-			case MSG_TRIGGER_STATE:{
-				if(peer){
-					triggerGraph.ApplySerializedState(data);
-				}
-			}break;
-			case MSG_CAMERA:{
-				Sint16 cx, cy;
-				memcpy(&cx, &data.data[data.BitsToBytes(data.readoffset)],     sizeof(cx));
-				memcpy(&cy, &data.data[data.BitsToBytes(data.readoffset) + 2], sizeof(cy));
-				if(cx != 0 || cy != 0){
-					pancameraactive = true;
-					pancamerareturn = false;
-					pancamerax = cx;
-					pancameray = cy;
-				} else {
-					pancameraactive = false;
-					pancamerareturn = true;
-					pancamerareturncount = GASLoader::Get().gameengine.ticksPerSecond * 3; // 3s return window
-				}
-			}break;
-		}
-	}
-	Uint32 tickcheck = SDL_GetTicks();
-	if(peerlist[authoritypeer]){
-		if(peerlist[authoritypeer]->lastpacket < tickcheck && tickcheck - peerlist[authoritypeer]->lastpacket >= peertimeout){
-			HandleDisconnect(authoritypeer);
-		}
-	}
-}
-
-Peer * World::AddPeer(char * address, unsigned short port, Uint8 agency, Uint32 accountid, bool observer){
-	Uint8 newpeerid = 0;
-	sockaddr_in addr;
-	addr.sin_addr.s_addr = inet_addr(address);
-	addr.sin_port = htons(port);
-	Peer * peer = FindPeer(addr);
-	if(!peer){
-		bool peeradded = false;
-		for(unsigned int i = 1; i < maxpeers; i++){
-			if(!peerlist[i]){
-				newpeerid = i;
-				peeradded = true;
-				break;
-			}
-		}
-		Peer * newpeer = new Peer();
-		newpeer->id = newpeerid;
-		newpeer->ip = ntohl(inet_addr(address));
-		newpeer->port = port;
-		newpeer->accountid = accountid;
-		if(peeradded){
-			if(!observer){
-				if(!FindTeamForPeer(*newpeer, agency)){
-					//printf("could not find team for new peer\n");
-					delete newpeer;
-					return 0;
-				}
-			}
-			peerlist[newpeerid] = newpeer;
-			peercount++;
-			//printf("new peer added, peer id: %d (%s:%d) peercount = %d\n", newpeerid, address, port, peercount);
-			return newpeer;
-		}else{
-			delete newpeer;
-			return 0;
-		}
-	}else{
-		//printf("existing peer added, peer id: %d\n", peer->id);
-		return 0;
-		//return peer;
-	}
-	return 0;
-}
-
-Peer * World::AddBot(Uint8 agency){
-	Uint8 newpeerid = 0;
-	bool peeradded = false;
-	for(unsigned int i = 1; i < maxpeers; i++){
-		if(!peerlist[i]){
-			newpeerid = i;
-			peeradded = true;
-			break;
-		}
-	}
-	Peer * newpeer = new Peer();
-	newpeer->id = newpeerid;
-	if(peeradded){
-		if(!FindTeamForPeer(*newpeer, agency)){
-			//printf("could not find team for new peer\n");
-			delete newpeer;
-			return 0;
-		}
-		newpeer->isbot = true;
-		peerlist[newpeerid] = newpeer;
-		peercount++;
-		//printf("added bot, peer id: %d peercount = %d\n", newpeerid, peercount);
-		return newpeer;
-	}else{
-		delete newpeer;
-		return 0;
-	}
-	return 0;
-}
-
-Peer * World::FindPeer(sockaddr_in & sockaddr){
-	Peer * peer = 0;
-	for(int i = 0; i < maxpeers; i++){
-		if(peerlist[i]){
-			if(peerlist[i]->ip == ntohl(sockaddr.sin_addr.s_addr)){
-				if(peerlist[i]->port == ntohs(sockaddr.sin_port)){
-					peer = peerlist[i];
-				}
-			}
-		}
-	}
-	return peer;
-}
-
-bool World::ProcessInputQueue(Peer & peer){
-	if(inputqueue[peer.id].size() > 0){
-		Serializer * data = inputqueue[peer.id].front();
-		Uint32 theirlasttick;
-		data->Get(theirlasttick);
-		Uint32 lasttick;
-		data->Get(lasttick);
-		if(lasttick >= peer.lasttick){
-			peer.theirlasttick = theirlasttick;
-			peer.lasttick = lasttick;
-			//peer.port = ntohs(senderaddr.sin_port);
-			peer.input.Serialize(Serializer::READ, *data);
-			for(std::list<Uint16>::iterator i = peer.controlledlist.begin(); i != peer.controlledlist.end(); i++){
-				Object * object = GetObjectFromId((*i));
-				if(object){
-					//if(!replaying){
-						object->oldx = object->x;
-						object->oldy = object->y;
-					//}
-					if (!input_locked) object->HandleInput(peer.input);
-					object->Tick(*this);
-					object->lasttick = tickcount;
-					//printf("Processed input for peer %d at tick %d\n", peer.id, tickcount);
-				}
-			}
-		}
-		delete data;
-		inputqueue[peer.id].pop_front();
-		return true;
-	}
-	return false;
-}
-
-void World::ProcessSnapshotQueue(void){
-	if(!peerlist[localpeerid]){
-		return;
-	}
-	int maxruns = 1;
-	if(snapshotqueue.size() > snapshotqueuemaxsize){
-		maxruns = 2;
-	}
-	if(snapshotqueue.size() == 0){
-		if(snapshotqueuemaxsize < GASLoader::Get().gameengine.snapshotQueueMaxCap){
-			snapshotqueuemaxsize++;
-		}
-	}
-	if(tickcount - lastsnapshotqueueadjust > (Uint32)GASLoader::Get().gameengine.snapshotQueueShrinkTicks){
-		if(snapshotqueuemaxsize - snapshotqueueminsize > 1){
-			snapshotqueuemaxsize--;
-			lastsnapshotqueueadjust = tickcount;
-		}
-	}
-	if(snapshotqueue.size() < snapshotqueueminsize){
-		return;
-	}
-	snapshotqueue.sort(CompareSnapshot);
-	for(int k = 0; k < maxruns; k++){
-		if(snapshotqueue.size() > 0){
-			Serializer * data = snapshotqueue.front();
-			unsigned int oldreadoffset = data->readoffset;
-			Uint32 tick;
-			data->Get(tick);
-			Uint32 ourtick;
-			data->Get(ourtick);
-			unsigned int readoffsetbeforedelta = data->readoffset;
-			Uint32 deltatick;
-			data->Get(deltatick);
-			data->readoffset = readoffsetbeforedelta;
-			Serializer ** deltasnapshotptr = &oldsnapshots[localpeerid][deltatick % maxoldsnapshots];
-			
-			if(DELTAENABLED && *deltasnapshotptr && tick - deltatick < maxoldsnapshots){
-				LoadSnapshot(*data, true, *deltasnapshotptr);
-			}else{
-				LoadSnapshot(*data, true);
-			}
-			
-			Serializer ** newsnapshotptr = &oldsnapshots[localpeerid][tick % maxoldsnapshots];
-			if(!*newsnapshotptr){
-				*newsnapshotptr = new Serializer;
-			}
-			(*newsnapshotptr)->offset = 0;
-			SaveSnapshot(**newsnapshotptr, localpeerid);
-			
-			TickObjects();
-			
-			if(tick > peerlist[localpeerid]->lasttick){
-				peerlist[localpeerid]->lasttick = tick;
-				ClientSidePredict(ourtick);
-			}
-
-			data->readoffset = oldreadoffset;
-			delete data;
-			snapshotqueue.pop_front();
-		}
-	}
-}
-
-void World::ClientSidePredict(Uint32 ourtick){
-	// Perform client side predication
-	int replayticks = tickcount - ourtick - 1;
-	if(replayticks < maxlocalinputhistory - 1){
-		for(int i = replayticks; i > 0; i--){
-			for(std::list<Uint16>::iterator it = peerlist[localpeerid]->controlledlist.begin(); it != peerlist[localpeerid]->controlledlist.end(); it++){
-				Object * object = GetObjectFromId((*it));
-				if(object){
-					// set the oldinput
-					if(object->type == ObjectTypes::PLAYER){
-						Player * player = (Player *)object;
-						player->oldinput = localinputhistory[(tickcount - i - 1) % maxlocalinputhistory];
-					}
-					//
-					object->HandleInput(localinputhistory[(tickcount - i) % maxlocalinputhistory]);
-					object->oldx = object->x;
-					object->oldy = object->y;
-					replaying = true;
-					audio.enabled = false;
-					object->Tick(*this);
-					object->lasttick = tickcount;
-					audio.enabled = true;
-					replaying = false;
-				}
-			}
-		}
-	}
-	//
-}
-
-void World::CheckExists(void){
-	Serializer data;
-	Uint8 code = MSG_EXISTS;
-	data.Put(code);
-	int count = 0;
-	Uint8 types[] = {ObjectTypes::PICKUP, ObjectTypes::FIXEDCANNON, ObjectTypes::DETONATOR};
-	// check for these objects, because they can get destroyed while we are not there, and they wont get deleted
-	for(int i = 0; i < sizeof(types) / sizeof(Uint8); i++){
-		for(std::vector<Uint16>::iterator it = objectsbytype[types[i]].begin(); it != objectsbytype[types[i]].end(); it++){
-			if(count >= GASLoader::Get().gameengine.maxStaleSnapshots){
-				break;
-			}
-			Object * object = GetObjectFromId(*it);
-			if(object && tickcount - object->lastsnapshottick > maxoldsnapshots){
-				data.Put((*it));
-				count++;
-			}
-		}
-	}
-	SendPacket(GetAuthorityPeer(), data.data, data.BitsToBytes(data.offset));
-}
-
-void World::ClearSnapshotQueue(void){
-	for(std::list<Serializer *>::iterator it = snapshotqueue.begin(); it != snapshotqueue.end(); it++){
-		Serializer * data = *it;
-		delete data;
-	}
-	snapshotqueue.clear();
-}
-
-void World::ClearMapData(void){
-	currentmapdata.clear();
-	currentmapdataprocessed = false;
-	currentmapdataend = false;
-}
-
-void World::AllocateMapData(int size){
-	ClearMapData();
-	currentmapdata.resize(size);
-}
-
-void World::ReadPeerList(Serializer & data){
-	for(unsigned int i = 0; i < maxpeers; i++){
-		if(i == authoritypeer){
-			continue;
-		}
-		Peer * peer = peerlist[i];
-		if(peer){
-			delete peer;
-			peerlist[i] = 0;
-			peercount--;
-		}
-	}
-	while(data.MoreBytesToRead()){
-		Peer * peer = new Peer();
-		peer->Serialize(Serializer::READ, data);
-		if(peer->id == authoritypeer && peerlist[peer->id]){
-			peerlist[peer->id]->accountid = peer->accountid;
-			peerlist[peer->id]->controlledlist = peer->controlledlist;
-			peerlist[peer->id]->gameinfoloaded = peer->gameinfoloaded;
-			peerlist[peer->id]->isready = peer->isready;
-			peerlist[peer->id]->ishost = peer->ishost;
-			delete peer;
-			continue;
-		}
-		peerlist[peer->id] = peer;
-		//printf("peerlist[%d]\n", peer->id);
-		peercount++;
-	}
-}
-
-void World::SendPacket(Peer * peer, char * data, unsigned int size){
-	if(replay.IsPlaying()){
-		return;
-	}
-	if(peer){
-		if(lagsimulator.Active() && gameplaystate == INGAME){
-			lagsimulator.QueuePacket(peer, data, size);
-		}else{
-			sockaddr_in recvaddr;
-			recvaddr.sin_family = AF_INET;
-			recvaddr.sin_port = htons(peer->port);
-			recvaddr.sin_addr.s_addr = htonl(peer->ip);
-			int ret = sendto(sockethandle, data, size, 0, (sockaddr *)&recvaddr, sizeof(recvaddr));
-			if(ret > 0){
-				totalbytessent += ret;
-			}
-		}
-	}
-}
-
-void World::SwitchToMode(bool newmode){
-	if(newmode == REPLICA && mode == AUTHORITY){
-		mode = REPLICA;
-		for(unsigned int i = 0; i < maxpeers; i++){
-			if(peerlist[i]){
-				authoritypeer = i;
-				break;
-			}
-		}
-	}else
-	if(newmode == AUTHORITY && mode == REPLICA){
-		mode = AUTHORITY;
-		authoritypeer = localpeerid;
-		for(unsigned int i = 0; i < maxpeers; i++){
-			if(peerlist[i]){
-				peerlist[i]->lastpacket = SDL_GetTicks();
-			}
-		}
-		Listen(GetAuthorityPeer()->port);
-		SendPeerList();
-	}
-}
-
-void World::DeleteOldSnapshots(Uint8 peerid){
-	for(unsigned int i = 0; i < maxoldsnapshots; i++){
-		if(oldsnapshots[peerid][i]){
-			delete oldsnapshots[peerid][i];
-			oldsnapshots[peerid][i] = 0;
-		}
-	}
-}
-
-void World::HandleDisconnect(Uint8 peerid, bool permanent){
-	//printf("peer %d disconnected\n", peerid);
-	if(replay.IsRecording()){
-		replay.WriteDisconnect(peerid);
-	}
-	bool park = (!permanent && mode == AUTHORITY && gameplaystate == INGAME && peerlist[peerid] && peerlist[peerid]->accountid != 0 && !peerlist[peerid]->isbot && !peerlist[peerid]->observer);
-	for(std::list<Uint16>::iterator i = peerlist[peerid]->controlledlist.begin(); i != peerlist[peerid]->controlledlist.end(); i++){
-		Object * object = GetObjectFromId((*i));
-		if(object){
-			object->HandleDisconnect(*this, peerid);
-		}
-	}
-	if(park){
-		peerlist[peerid]->disconnected = true;
-		peerlist[peerid]->isready = false;
-		SendPeerList();
-		return;
-	}
-	ClearSnapshotQueue();
-	// Capture team before RemovePeer strips the peer from all teams.
-	// GetPeerTeam searches team membership lists, so it must run first.
-	Team * leavingTeam = (mode == AUTHORITY) ? GetPeerTeam(peerid) : 0;
-	for(std::list<Object *>::iterator it = objectlist.begin(); it != objectlist.end(); it++){
-		Object * object = *it;
-		if(object->type == ObjectTypes::TEAM){
-			Team * team = static_cast<Team *>(object);
-			if(team){
-				team->RemovePeer(peerid);
-			}
-		}
-	}
-	if(mode == REPLICA){
-		delete peerlist[peerid];
-		peerlist[peerid] = 0;
-		peercount--;
-		if(peerid == authoritypeer){
-			ShowMessage("CONNECTION LOST", 128, 20);
-			SwitchToLocalAuthorityMode();
-			state = IDLE;
-		}
-	}else
-	if(mode == AUTHORITY){
-		// If a player disconnects mid-game, record their partial stats immediately
-		// so their progress isn't lost (won=0 since they left before the game ended).
-		if(gameplaystate == INGAME && peerlist[peerid]->accountid != 0){
-			Peer * leavingPeer = peerlist[peerid];
-			User * user = lobby.GetUserInfo(leavingPeer->accountid);
-			if(user && leavingTeam){
-				user->statscopy = leavingPeer->stats;
-				user->statsagency = leavingTeam->agency;
-				user->teamnumber = leavingTeam->number;
-				lobby.RegisterStats(*user, 0, gameinfo.id);
-			}
-		}
-		delete peerlist[peerid];
-		peerlist[peerid] = 0;
-		peercount--;
-		for(int i = 0; i < maxoldsnapshots; i++){
-			if(oldsnapshots[peerid][i]){
-				delete oldsnapshots[peerid][i];
-				oldsnapshots[peerid][i] = 0;
-			}
-		}
-		SendPeerList();
-	}
-}
-
-bool World::RelevantToPlayer(Player * player, Object * object){
-	switch(object->type){
-		case ObjectTypes::TEAM:
-		case ObjectTypes::STATE:
-			return true;
-		break;
-		case ObjectTypes::PLAYER:{
-			Player * objplayer = static_cast<Player *>(object);
-			if(objplayer->hassecret || objplayer->state == Player::UNDEPLOYING){
-				return true;
-			}
-		}break;
-		case ObjectTypes::PICKUP:{
-			PickUp * pickup = static_cast<PickUp *>(object);
-			if(pickup->type == PickUp::SECRET){
-				return true;
-			}
-		}break;
-		case ObjectTypes::SURVEILLANCEMONITOR:{
-			
-		}break;
-	}
-	if(!player){
-		return false;
-	}
-	if(rand() % objectlist.size() == 0){
-		return true;
-	}
-	if(object->snapshotinterval >= 0 && tickcount % (object->snapshotinterval + 1) == 0){
-		return true;
-	}
-	if(object->issprite){
-		const WorldDef& _wd = GASLoader::Get().world;
-		if(abs(player->x - object->x) <= _wd.networkSyncRangeX && abs(player->y - object->y) <= _wd.networkSyncRangeY){
-			return true;
-		}
-		for(std::vector<Uint16>::iterator it = objectsbytype[ObjectTypes::SURVEILLANCEMONITOR].begin(); it != objectsbytype[ObjectTypes::SURVEILLANCEMONITOR].end(); it++){
-			Object * obj = GetObjectFromId((*it));
-			if(obj){
-				if(abs(player->x - obj->x) <= _wd.networkSyncRangeX && abs(player->y - obj->y) <= _wd.networkSyncRangeY){
-					SurveillanceMonitor * surveillancemonitor = static_cast<SurveillanceMonitor *>(obj);
-					if(surveillancemonitor->camera.IsVisible(*this, *object)){
-						return true;
-					}
-				}
-			}
-		}
-		Object * grenade = GetObjectFromId(player->currentgrenade);
-		if(grenade){
-			if(abs(grenade->x - object->x) <= _wd.grenadesyncRangeX && abs(grenade->y - object->y) <= _wd.grenadesyncRangeY){
-				return true;
-			}
-		}
-		Object * detonator = GetObjectFromId(player->currentdetonator);
-		if(detonator){
-			if(abs(detonator->x - object->x) <= _wd.grenadesyncRangeX && abs(detonator->y - object->y) <= _wd.grenadesyncRangeY){
-				return true;
-			}
-		}
-	}
-	return false;
-}
-
-bool World::BelongsToTeam(Object & object, Uint16 teamid){
-	switch(object.type){
-		case ObjectTypes::PLAYER:{
-			Player * player = static_cast<Player *>(&object);
-			if(player->teamid == teamid){
-				return true;
-			}
-		}break;
-		case ObjectTypes::CIVILIAN:{
-			Civilian * civilian = static_cast<Civilian *>(&object);
-			if(civilian->tractteamid == teamid){
-				return true;
-			}
-		}break;
-		case ObjectTypes::ROBOT:{
-			Robot * robot = static_cast<Robot *>(&object);
-			if(robot->virusplanter == teamid){
-				return true;
-			}
-		}break;
-		case ObjectTypes::FIXEDCANNON:{
-			FixedCannon * fixedcannon = static_cast<FixedCannon *>(&object);
-			Player * player = static_cast<Player *>(GetObjectFromId(fixedcannon->ownerid));
-			if(player && player->teamid == teamid){
-				return true;
-			}
-		}break;
-		case ObjectTypes::WALLDEFENSE:{
-			WallDefense * walldefense = static_cast<WallDefense *>(&object);
-			if(walldefense->teamid == teamid){
-				return true;
-			}
-		}break;
-		case ObjectTypes::TECHSTATION:{
-			TechStation * techstation = static_cast<TechStation *>(&object);
-			if(techstation->teamid == teamid){
-				return true;
-			}
-		}break;
-	}
-	return false;
-}
-
-void World::ActivateTerminals(void){
-	std::vector<Terminal *> terminallist;
-	for(std::list<Object *>::iterator it = objectlist.begin(); it != objectlist.end(); it++){
-		Object * object = *it;
-		if(object->type == ObjectTypes::TERMINAL){
-			Terminal * terminal = static_cast<Terminal *>(object);
-			if(terminal){
-				terminallist.push_back(terminal);
-			}
-		}
-	}
-	//std::random_shuffle(terminallist.begin(), terminallist.end());
-	for(int i = 0; i < terminallist.size(); i++){
-		int r = Random() % terminallist.size();
-		Terminal * temp = terminallist[i];
-		terminallist[i] = terminallist[r];
-		terminallist[r] = temp;
-	}
-	
-	int numused = 0;
-	int numsecret = 0;
-	for(int i = 0; i < terminallist.size(); i++){
-		switch(terminallist[i]->state){
-			case Terminal::HACKING:
-			case Terminal::BEAMING:
-			case Terminal::READY:{
-				numused++;
-			}break;
-			case Terminal::SECRETBEAMING:
-			case Terminal::SECRETREADY:{
-				numsecret++;
-			}
-		}
-	}
-	const WorldDef& _wdef = GASLoader::Get().world;
-	int numtoactivate = (int)(terminallist.size() * _wdef.terminalActivatePercent);
-	numtoactivate -= numused;
-	numtoactivate += numsecret;
-	int numactivated = 0;
-	if(numtoactivate > 0){
-		for(int i = 0; i < (int)terminallist.size(); i++){
-			Terminal * terminal = terminallist[i];
-			if(terminal->state == Terminal::INACTIVE){
-				terminal->state = Terminal::BEAMING;
-				if(terminal->isbig){
-					terminal->beamingseconds = (Random() % _wdef.terminalBigBeamRange) + _wdef.terminalBigBeamMin;
-				}else{
-					terminal->beamingseconds = (Random() % _wdef.terminalSmallBeamRange) + _wdef.terminalSmallBeamMin;
-				}
-				numactivated++;
-				if(numactivated >= numtoactivate){
-					break;
-				}
-			}
-		}
-	}
-}
-
-void World::LoadBuyableItems(void){
-	for(const ItemDef& def : GASLoader::Get().items){
-		buyableitems.push_back(new BuyableItem(
-			(Uint8)def.enumId,
-			def.name.c_str(),
-			def.price,
-			def.repairPrice,
-			(Uint8)def.spriteBank,
-			(Uint8)def.spriteIndex,
-			(Uint32)def.techChoice,
-			(Uint8)def.techSlots,
-			def.description.c_str(),
-			def.agencyRestriction
-		));
-	}
-}
-
-void World::BuyItem(Uint8 id){
-	char msg[3];
-	msg[0] = MSG_STATION;
-	msg[1] = STA_BUY;
-	msg[2] = id;
-	SendPacket(GetAuthorityPeer(), msg, sizeof(msg));
-}
-
-void World::RepairItem(Uint8 id){
-	char msg[3];
-	msg[0] = MSG_STATION;
-	msg[1] = STA_REPAIR;
-	msg[2] = id;
-	SendPacket(GetAuthorityPeer(), msg, sizeof(msg));
-}
-
-void World::VirusItem(Uint8 id){
-	char msg[3];
-	msg[0] = MSG_STATION;
-	msg[1] = STA_VIRUS;
-	msg[2] = id;
-	SendPacket(GetAuthorityPeer(), msg, sizeof(msg));
-}
-
-void World::ChangeTeam(Uint8 peerid){
-	Peer * peer = peerlist[peerid];
-	if(peer){
-		Team * peerteam = GetPeerTeam(peer->id);
-		if(peerteam){
-			int start = peerteam->number + 1;
-			if(FindTeamForPeer(*peer, peerteam->agency, start)){
-				peerteam->RemovePeer(peer->id);
-				peer->isready = false;
-				SendPeerList();
-			}
-		}
-	}
-}
-
-void World::SetTech(Uint8 peerid, Uint32 techchoices){
-	Peer * peer = peerlist[peerid];
-	if(peer){
-		lobby.GetUserInfo(peer->accountid);
-		peer->wantedtechchoices = techchoices;
-		ApplyWantedTech(*peer);
-	}
-}
-
-void World::DisplayChatMessage(Uint32 accountid, const char * msg){
-	std::string chatmsg(lobby.GetUserInfo(accountid)->name);
-	std::replace(chatmsg.begin(), chatmsg.end(), ' ', '\xA0');
-	// replace spaces with nbsp so usernames dont wordwrap
-	chatmsg.append(":\xA0");
-	chatmsg.append(msg);
-	
-	char * wrapped = silencer::ui::WordWrapText(chatmsg.c_str(), 36, "\n ");
-	char * line = strtok(wrapped, "\n");
-	while(line){
-		chatlines.push_back(line);
-		line = strtok(NULL, "\n");
-	}
-	delete[] wrapped;
-	
-	showchat_i = GASLoader::Get().gameengine.chatDisplayTicks;
-	while((int)chatlines.size() > GASLoader::Get().gameengine.chatMaxLines){
-		chatlines.pop_front();
-	}
-}
-
-void World::SendStats(Peer & peer){
-	Serializer msg;
-	Uint8 code = MSG_STATS;
-	msg.Put(code);
-	peer.stats.Serialize(Serializer::WRITE, msg);
-	SendPacket(&peer, msg.data, msg.BitsToBytes(msg.offset));
-}
-
-void World::UserInfoReceived(Peer & peer){
-	if(replay.IsRecording()){
-		replay.WriteUserInfo(*lobby.GetUserInfo(peer.accountid));
-	}
-	ApplyWantedTech(peer);
-}
-
-bool World::IsCollidable(Uint8 type){
-	switch(type){
-		case ObjectTypes::SHRAPNEL:
-		case ObjectTypes::PLUME:
-			return false;
-		break;
-		default:
-			return true;
-		break;
-	}
-}
-
-void World::ApplyWantedTech(Peer & peer){
-	Team * peerteam = GetPeerTeam(peer.id);
-	User * user = lobby.GetUserInfo(peer.accountid);
-	Uint32 techchoices = peer.wantedtechchoices;
-	if(peerteam){
-		Uint32 oldtechchoices = peer.techchoices;
-		for(std::vector<BuyableItem *>::iterator it = buyableitems.begin(); it != buyableitems.end(); it++){
-			BuyableItem * buyableitem = *it;
-			if(buyableitem->agencyspecific != -1 && buyableitem->agencyspecific != peerteam->agency){
-				techchoices &= ~(buyableitem->techchoice);
-			}
-		}
-		peer.techchoices = techchoices;
-		if(user && user->agency[peerteam->agency].techslots >= TechSlotsUsed(peer)){
-			SendPeerList();
-		}else{
-			peer.techchoices = oldtechchoices;
-		}
-	}
-}
-
-bool World::CompareTeamByNumber(Team * team1, Team * team2){
-	return(team1->number < team2->number);
-}
-
-bool World::CompareSnapshot(Serializer * snapshot1, Serializer * snapshot2){
-	unsigned int oldreadoffset = snapshot1->readoffset;
-	Uint32 tick1;
-	snapshot1->Get(tick1);
-	snapshot1->readoffset = oldreadoffset;
-	oldreadoffset = snapshot2->readoffset;
-	Uint32 tick2;
-	snapshot2->Get(tick2);
-	snapshot2->readoffset = oldreadoffset;
-	return(tick1 < tick2);
-}
-
-bool World::Listen(unsigned short port){
-	if(!boundport){
-		if(!Bind(port)){
-			return false;
-		}
-	}
-	AllocateMapData(65535);
-	SwitchToMode(AUTHORITY);
-	state = LISTENING;
-	Peer * authoritypeerptr = GetAuthorityPeer();
-	authoritypeerptr->ip = INADDR_ANY;
-	authoritypeerptr->port = boundport;
-	//printf("Listening on port %d\n", boundport);
-	return true;
-}
-
-unsigned short World::Bind(unsigned short port){
-	sockaddr_in recvaddr;
-	recvaddr.sin_family = AF_INET;
-	recvaddr.sin_port = htons(port);
-	recvaddr.sin_addr.s_addr = INADDR_ANY;
-	int ret = bind(sockethandle, (sockaddr *)&recvaddr, sizeof(recvaddr));
-	if(ret == 0){
-		sockaddr_in boundaddr;
-		socklen_t boundaddrlen = sizeof(boundaddr);
-		getsockname(sockethandle, (sockaddr *)&boundaddr, &boundaddrlen);
-		boundport = ntohs(boundaddr.sin_port);
-		return boundport;
-	}
-	return false;
-}
-
-void World::Connect(Uint8 agency, Uint32 accountid, const char * password, bool observer){
-	AllocateMapData(65535);
-	sockaddr_in addr;
-	addr.sin_addr.s_addr = htonl(GetAuthorityPeer()->ip);
-	//printf("sending connect request with agency %d, accountid %d to %s:%d\n", agency, accountid, inet_ntoa(addr.sin_addr), GetAuthorityPeer()->port);
-	SwitchToMode(REPLICA);
-	state = CONNECTING;
-	messagetype = 0;
-	authoritypeer = 0;
-	GetAuthorityPeer()->lastpacket = SDL_GetTicks();
-	Serializer data;
-	Uint8 code = MSG_CONNECT;
-	data.Put(code);
-	data.Put(agency);
-	data.Put(accountid);
-	Uint8 passwordsize = password ? strlen(password) : 0;
-	data.Put(passwordsize);
-	for(int i = 0; i < passwordsize; i++){
-		data.Put(password[i]);
-	}
-	data.PutBit(observer);
-	SendPacket(GetAuthorityPeer(), data.data, data.BitsToBytes(data.offset));
-}
-
-void World::Disconnect(void){
-	ClearSnapshotQueue();
-	ClearMapData();
-	state = IDLE;
-	viewedpeerid = 0;
-	spectator.freecam = false;
-	spectator.camx = 0;
-	spectator.camy = 0;
-	spectator.camvx = 0;
-	spectator.camvy = 0;
-	spectator.holdshowallnames = false;
-	spectator.initialized = false;
-	char data[1];
-	data[0] = MSG_DISCONNECT;
-	if(mode == AUTHORITY){
-		for(int i = 0; i < maxpeers; i++){
-			Peer * peer = peerlist[i];
-			if(peer){
-				SendPacket(peer, data, sizeof(data));
-			}
-		}
-	}else
-	if(mode == REPLICA){
-		SendPacket(GetAuthorityPeer(), data, sizeof(data));
-		HandleDisconnect(GetAuthorityPeer()->id);
-	}
-}
-
-void World::SendInput(void){
-	Peer * peer = 0;
-	localinputhistory[tickcount % maxlocalinputhistory] = localinput;
-	if(mode == REPLICA && state == CONNECTED && gameplaystate == INGAME){
-		peer = peerlist[localpeerid];
-		if(peer && peer->controlledlist.size() > 0){
-			peer->input = localinput;
-			peer->input.mousex = 0xFFFF;
-			peer->input.mousey = 0xFFFF;
-			Serializer data;
-			Uint8 code = MSG_INPUT;
-			data.Put(code);
-			data.Put(tickcount);
-			data.Put(peerlist[localpeerid]->lasttick);
-			peer->input.Serialize(Serializer::WRITE, data);
-			SendPacket(GetAuthorityPeer(), data.data, data.BitsToBytes(data.offset));
-		}
-	}else
-	if(mode == AUTHORITY){
-		peer = GetAuthorityPeer();
-		peer->input = localinput;
-	}
-	if(peer){
-		for(std::list<Uint16>::iterator i = peer->controlledlist.begin(); i != peer->controlledlist.end(); i++){
-			Object * object = GetObjectFromId((*i));
-			if(object){
-				//if(!replaying){
-					object->oldx = object->x;
-					object->oldy = object->y;
-				//}
-				if (!input_locked) object->HandleInput(peer->input);
-				object->Tick(*this);
-				object->lasttick = tickcount;
-			}
-		}
-	}
-}
-
-Peer * World::GetAuthorityPeer(void){
-	if(!peerlist[authoritypeer]){
-		peerlist[authoritypeer] = new Peer();
-		peerlist[authoritypeer]->ip = INADDR_ANY;
-		peerlist[authoritypeer]->id = authoritypeer;
-	}
-	return peerlist[authoritypeer];
-}
-
-Peer * World::GetPeer(Uint8 peerid){
-	if(peerid >= maxpeers) return 0;
-	return peerlist[peerid];
-}
-
-Player * World::GetPeerPlayer(Uint8 peerid){
-	Object * object = 0;
-	Player * player = 0;
-	if(peerlist[peerid]){
-		if(peerlist[peerid]->controlledlist.size() > 0){
-			object = GetObjectFromId((*peerlist[peerid]->controlledlist.begin()));
-			if(object && object->type == ObjectTypes::PLAYER){
-				player = static_cast<Player *>(object);
-				if(player){
-					return player;
-				}
-			}
-		}
-	}
-	return 0;
-}
-
-Team * World::GetPeerTeam(Uint8 peerid){
-	if(peerlist[peerid]){
-		for(std::vector<Uint16>::iterator it = objectsbytype[ObjectTypes::TEAM].begin(); it != objectsbytype[ObjectTypes::TEAM].end(); it++){
-			Team * team = static_cast<Team *>(GetObjectFromId((*it)));
-			for(int i = 0; i < team->numpeers; i++){
-				if(team->peers[i] == peerid){
-					return team;
-				}
-			}
-		}
-	}
-	return 0;
-}
-
-bool World::FindTeamForPeer(Peer & peer, Uint8 agency, int start){
-	int maxteams2 = maxteams;
-	if(dedicatedserver.active){
-		maxteams2 = gameinfo.maxteams;
-		if(maxteams2 > maxteams){
-			maxteams2 = maxteams;
-		}
-	}
-	if(start >= maxteams2){
-		start = 0;
-	}
-	int teamnumber = start;
-	bool teamfound = false;
-	bool slotfound = true;
-	std::vector<Team *> teamlist;
-	for(std::list<Object *>::iterator it = objectlist.begin(); it != objectlist.end(); it++){
-		Object * object = (*it);
-		if(object->type == ObjectTypes::TEAM){
-			Team * team = static_cast<Team *>(object);
-			if(team){
-				teamlist.push_back(team);
-			}
-		}
-	}
-	std::sort(teamlist.begin(), teamlist.end(), CompareTeamByNumber);
-	std::vector<Team *>::iterator it = teamlist.begin();
-	while(it != teamlist.end()){
-		Team * team = *it;
-		if(team->number == start){
-			slotfound = false;
-		}
-		if(team->number == teamnumber){
-			teamnumber = team->number + 1;
-			it = teamlist.begin();
-		}else{
-			it++;
-		}
-	}
-	if(!slotfound){
-		for(std::vector<Team *>::iterator it = teamlist.begin(); it != teamlist.end(); it++){
-			Team * team = *it;
-			if(team->number >= start){
-				if(team->agency == agency){
-					if(team->AddPeer(peer.id)){
-						teamfound = true;
-						break;
-					}
-				}
-			}
-		}
-	}
-	if(!teamfound){
-		if(teamlist.size() >= maxteams2){
-			return false;
-		}
-		Team * newteam = (Team *)CreateObject(ObjectTypes::TEAM);
-		newteam->agency = agency;
-		newteam->number = teamnumber;
-		newteam->AddPeer(peer.id);
-	}
-	return true;
-}
-
-void World::SendSnapshots(void){
-	for(unsigned int i = 0; i < maxpeers; i++){
-		Peer * peer = peerlist[i];
-		if(peer && i != localpeerid && !peer->isbot && !peer->disconnected){
-			Serializer data;
-			Uint8 code = MSG_SNAPSHOT;
-			data.Put(code);
-			data.Put(tickcount);
-			data.Put(peer->theirlasttick);
-			SaveSnapshot(data, i);
-			SendPacket(peer, data.data, data.BitsToBytes(data.offset));
-		}
-	}
-}
-
-void World::SendGameInfo(Uint8 peerid){
-	Peer * peer = peerlist[peerid];
-	if(peer){
-		Serializer data;
-		Uint8 code = MSG_GAMEINFO;
-		data.Put(code);
-		gameinfo.Serialize(Serializer::WRITE, data);
-		SendPacket(peer, data.data, data.BitsToBytes(data.offset));
-	}
-}
-
-void World::SendGameInfoLoaded(void){
-	char data[1];
-	data[0] = MSG_GAMEINFO;
-	SendPacket(GetAuthorityPeer(), data, sizeof(data));
-}
-
-void World::SendReady(void){
-	Serializer data;
-	Uint8 code = MSG_READY;
-	data.Put(code);
-	SendPacket(GetAuthorityPeer(), data.data, data.BitsToBytes(data.offset));
-}
-
-bool World::AllPeersReady(Uint8 except){
-	bool allready = true;
-	for(int i = 0; i < maxpeers; i++){
-		Peer * peer = peerlist[i];
-		if(peer){
-			if(!peer->isready && peer->id != except){
-				allready = false;
-				break;
-			}
-		}
-	}
-	return allready;
-}
-
-bool World::AllPeersLoadedGameInfo(void){
-	bool allloaded = true;
-	for(int i = 0; i < maxpeers; i++){
-		Peer * peer = peerlist[i];
-		if(peer){
-			if(!peer->gameinfoloaded){
-				allloaded = false;
-				break;
-			}
-		}
-	}
-	return allloaded;
-}
-
-bool World::AllPeersDownloadedMap(void){
-	bool allloaded = true;
-	for(int i = 0; i < maxpeers; i++){
-		if(i == authoritypeer) continue; // dedicated server doesn't download maps
-		Peer * peer = peerlist[i];
-		if(peer){
-			if(!peer->mapdownloaded){
-				allloaded = false;
-				break;
-			}
-		}
-	}
-	return allloaded;
-}
-
-void World::RequestPeerList(void){
-	if(authoritypeer != localpeerid){
-		Serializer data;
-		Uint8 code = MSG_PEERLIST;
-		data.Put(code);
-		SendPacket(GetAuthorityPeer(), data.data, data.BitsToBytes(data.offset));
-	}
-}
-
-void World::SendPeerList(Uint8 peerid){
-	if(mode == AUTHORITY){
-		Serializer data;
-		Uint8 code = MSG_PEERLIST;
-		data.Put(code);
-		for(unsigned int i = 0; i < maxpeers; i++){
-			Peer * peer = peerlist[i];
-			if(peer){
-				peer->Serialize(Serializer::WRITE, data);
-			}
-		}
-		for(unsigned int i = 0; i < maxpeers; i++){
-			Peer * peer = peerlist[i];
-			if(peer && i != localpeerid && (!peerid || peerid == peer->id) && !peer->disconnected){
-				SendPacket(peer, data.data, data.BitsToBytes(data.offset));
-			}
-		}
-	}
-}
-
-void World::SwitchToLocalAuthorityMode(void){
-	mode = AUTHORITY;
-	for(int i = 0; i < maxpeers; i++){
-		if(peerlist[i]){
-			delete peerlist[i];
-			peerlist[i] = 0;
-		}
-	}
-	peercount = 0;
-	authoritypeer = GetAuthorityPeer()->id;
-	localpeerid = authoritypeer;
-	viewedpeerid = localpeerid;
-}
-
-bool World::IsAuthority(void){
-	return mode == AUTHORITY;
-}
-
-bool World::IsLocalObserver(void){
-	Peer * lp = peerlist[localpeerid];
-	return lp && lp->observer;
-}
-
-bool World::IsConnected() const {
-	return state == CONNECTED;
-}
-
-bool World::IsIdle() const {
-	return state == IDLE;
-}
-
-void World::Illuminate(void){
-	illuminate = GASLoader::Get().world.illuminateLevel;
-}
-
-void World::ShowMessage(const char * message, Uint8 time, Uint8 type, bool networked, Peer * peer){
-	if(networked && IsAuthority()){
-		int msgsize = 1 + strlen(message) + 1 + 1 + 1;
-		char * msg = new char[msgsize];
-		msg[0] = MSG_MESSAGE;
-		memcpy(&msg[1], message, strlen(message) + 1);
-		msg[1 + strlen(message) + 1] = time;
-		msg[1 + strlen(message) + 1 + 1] = type;
-		if(!peer){
-			for(unsigned int i = 0; i < maxpeers; i++){
-				Peer * peer = peerlist[i];
-				if(peer && i != localpeerid){
-					SendPacket(peer, msg, msgsize);
-				}
-			}
-		}else{
-			SendPacket(peer, msg, msgsize);
-		}
-		delete[] msg;
-	}
-	if(!networked || (IsAuthority() && !peer) || (IsAuthority() && peer && peer->id == localpeerid)){
-		if(messagetype >= 10){ // Skip any messages after end of game messages, so it is not replaced
-			return;
-		}
-		strncpy(World::message, message, sizeof(World::message) - 1);
-		World::message[sizeof(World::message) - 1] = 0;
-		message_i = 1;
-		messagetime = time;
-		messagetype = type;
-	}
-}
-
-void World::ShowStatus(const char * status, Uint8 color, bool networked, Peer * peer){
-	char * newstatus = CreateStatusString(status, color, 100);
-	if(networked && IsAuthority()){
-		int msgsize = 1 + strlen(status) + 1 + 1 + 1;
-		char * msg = new char[msgsize];
-		msg[0] = MSG_STATUS;
-		memcpy(&msg[1], newstatus, strlen(status) + 1 + 1 + 1);
-		if(!peer){
-			for(unsigned int i = 0; i < maxpeers; i++){
-				Peer * peer = peerlist[i];
-				if(peer && i != localpeerid){
-					SendPacket(peer, msg, msgsize);
-				}
-			}
-		}else{
-			SendPacket(peer, msg, msgsize);
-		}
-		delete[] msg;
-	}
-	if(!networked || (IsAuthority() && !peer) || (IsAuthority() && peer && peer->id == localpeerid)){
-		PushStatusString(newstatus);
-	}
-}
-
-void World::ShowTopMessage(const char * message){
-	if(gameplaystate == INGAME){
-		memset(topmessage, 0, sizeof(topmessage));
-		strncpy(topmessage, message, sizeof(topmessage) - 1);
-		topmessage_i = 1;
-	}
-}
-
-void World::SendChat(bool toteam, char * message){
-	char msg[2 + 100 + 1];
-	memset(msg, 0, sizeof(msg));
-	msg[0] = MSG_CHAT;
-	msg[1] = toteam ? 1 : 0;
-	strncpy(&msg[2], message, 100);
-	msg[2 + 100] = 0;
-	SendPacket(GetAuthorityPeer(), msg, sizeof(msg));
-}
-
-void World::SendSound(const char * name, Peer * peer, Uint8 volume){
-	if(IsAuthority()){
-		if(!peer || (peer && peer->id == localpeerid)){
-			Audio::GetInstance().Play(resources.soundbank[name], volume);
-		}
-		char msg[2 + 255];
-		msg[0] = MSG_SOUND;
-		msg[1] = volume;
-		strcpy(&msg[2], name);
-		msg[2 + strlen(name)] = 0;
-		if(!peer){
-			for(unsigned int i = 0; i < maxpeers; i++){
-				Peer * peer = peerlist[i];
-				if(peer && i != localpeerid){
-					SendPacket(peer, msg, 2 + strlen(name) + 1);
-				}
-			}
-		}else{
-			if(peer->id != localpeerid){
-				SendPacket(peer, msg, 2 + strlen(name) + 1);
-			}
-		}
-	}
-}
-
-void World::BroadcastTriggerState() {
-    if (!IsAuthority()) return;
-    Serializer payload;
-    triggerGraph.SerializeState(payload);
-    int msgsize = 1 + payload.offset;
-    char * msg = new char[msgsize];
-    msg[0] = MSG_TRIGGER_STATE;
-    memcpy(&msg[1], payload.data, payload.offset);
-    for (unsigned int i = 0; i < maxpeers; i++) {
-        if (peerlist[i] && i != localpeerid) {
-            SendPacket(peerlist[i], msg, msgsize);
-        }
-    }
-    delete[] msg;
-}
-
-char * World::CreateStatusString(const char * status, Uint8 color, Uint8 duration){
-	char * newstatus = new char[strlen(status) + 1 + 1 + 1];
-	strcpy(newstatus, status);
-	newstatus[strlen(status) + 1] = duration;
-	newstatus[strlen(status) + 2] = color;
-	return newstatus;
-}
-
-void World::PushStatusString(char * statusstring){
-	while(statusmessages.size() >= maxstatusmessages){
-		delete[] statusmessages.back();
-		statusmessages.pop_back();
-	}
-	statusmessages.push_front(statusstring);
-}
-
-void World::ChangeTeam(void){
-	char msg[1];
-	msg[0] = MSG_CHANGETEAM;
-	SendPacket(GetAuthorityPeer(), msg, sizeof(msg));
-}
-
-void World::SetAgency(Uint8 agency){
-	char msg[2];
-	msg[0] = MSG_SETAGENCY;
-	msg[1] = agency;
-	SendPacket(GetAuthorityPeer(), msg, sizeof(msg));
-}
-
-void World::KillByGovt(Peer & peer){
-	if(IsAuthority()){
-		char msg[2];
-		msg[0] = MSG_GOVTKILL;
-		msg[1] = peer.id;
-		for(unsigned int i = 0; i < maxpeers; i++){
-			Peer * ipeer = peerlist[i];
-			if(ipeer){
-				SendPacket(ipeer, msg, 2);
-				if(ipeer->id == peer.id){
-					Player * player = GetPeerPlayer(ipeer->id);
-					if(player){
-						player->KillByGovt(*this);
-					}
-				}
-			}
-		}
-	}
-}
-
-bool World::IsSecurity(Object & object){
-	if(object.type == ObjectTypes::WALLDEFENSE){
-		WallDefense & walldefense = static_cast<WallDefense &>(object);
-		if(!walldefense.teamid){
-			return true;
-		}
-	}else
-	if(object.type == ObjectTypes::ROBOT){
-		Robot & robot = static_cast<Robot &>(object);
-		if(!robot.virusplanter){
-			return true;
-		}
-	}else
-	if(object.type == ObjectTypes::GUARD){
-		return true;
-	}
-	return false;
-}
-
-void World::Explode(Object & object, Uint8 suitcolor, float hitx){
-	object.draw = false;
-	for(int i = 0; i < 6; i++){
-		BodyPart * bodypart = (BodyPart *)CreateObject(ObjectTypes::BODYPART);
-		if(bodypart){
-			bodypart->suitcolor = suitcolor;
-			bodypart->x = object.x;
-			bodypart->y = object.y - GASLoader::Get().world.bodyPartSpawnYOffset;
-			bodypart->type = i;
-			bodypart->xv += (abs(object.xv) * 2) * hitx;
-			if(i == 0){
-				bodypart->xv = 0;
-				bodypart->yv = -GASLoader::Get().world.bodyPartLaunchYV;
-			}
-		}
-	}
-}
-
-void World::SetRandomSeed(Uint32 seed){
-	randomseed = seed;
-}
-
-Uint32 World::Random(void){
-	randomseed = 69069 * randomseed + 1;
-	return randomseed & 0x7FFF;
-}
-
-void World::SetTech(Uint32 techchoices){
-	Serializer data;
-	Uint8 code = MSG_TECH;
-	data.Put(code);
-	data.Put(techchoices);
-	//printf("MSG_TECH %d\n", techchoices);
-	SendPacket(GetAuthorityPeer(), data.data, data.BitsToBytes(data.offset));
-}
-
-int World::TechSlotsUsed(Peer & peer){
-	int slotsused = 0;
-	for(std::vector<BuyableItem *>::iterator it = buyableitems.begin(); it != buyableitems.end(); it++){
-		BuyableItem * buyableitem = *it;
-		if(buyableitem->techchoice & peer.techchoices){
-			slotsused += buyableitem->techslots;
-		}
-	}
-	return slotsused;
-}
-
-void World::SendMapDownloaded(void){
-	if(IsAuthority()){
-		peerlist[localpeerid]->mapdownloaded = true;
-		SendPeerList();
-	}else{
-		char data[2];
-		data[0] = MSG_MAP;
-		data[1] = MAP_DOWNLOADED;
-		SendPacket(GetAuthorityPeer(), data, sizeof(data));
-	}
-	//printf("sent MAP_DOWNLOADED\n");
-}
-
-void World::PutMapChunk(Uint32 offset, Peer & peer){
-	const Uint32 maxchunksize = 1024;
-	Uint32 size = maxchunksize;
-	if(size + offset > currentmapdata.size()){
-		if(offset >= currentmapdata.size()){
-			size = 0;
-		}else{
-			size = currentmapdata.size() - offset;
-		}
-	}
-	char data[1 + 1 + sizeof(Uint32) + sizeof(Uint32) + maxchunksize];
-	data[0] = MSG_MAP;
-	data[1] = MAP_PUTCHUNK;
-	*(Uint32 *)(&data[2]) = offset;
-	*(Uint32 *)(&data[2 + sizeof(Uint32)]) = size;
-	memcpy(&data[2 + sizeof(Uint32) + sizeof(Uint32)], &currentmapdata[offset], size);
-	SendPacket(&peer, data, sizeof(data));
-	//printf("sent MAP_PUTCHUNK %d %d\n", offset, size);
-}
-
-void World::GetMapChunk(Uint32 offset){
-	char data[1 + 1 + sizeof(Uint32)];
-	data[0] = MSG_MAP;
-	data[1] = MAP_GETCHUNK;
-	*(Uint32 *)(&data[2]) = offset;
-	if(IsAuthority()){
-		// get map chunk from host peer who has it already
-		for(int i = 0; i < maxpeers; i++){
-			Peer * peer = peerlist[i];
-			if(peer && peer->mapdownloaded){
-				SendPacket(peer, data, sizeof(data));
-				break;
-			}
-		}
-	}else{
-		SendPacket(GetAuthorityPeer(), data, sizeof(data));
-	}
-	//printf("sent MAP_GETCHUNK %d\n", offset);
-}
-
-void World::StoreMapChunk(unsigned char * data, Uint32 offset, Uint32 size){
-	//printf("StoreMapChunk %d %d\n", offset, size);
-	if(size == 0){
-		currentmapdataend = true;
-	}
-	if(offset + 1 > currentmapdata.size()){
-		return;
-	}
-	if(size + offset > currentmapdata.size()){
-		if(offset >= currentmapdata.size()){
-			size = 0;
-		}else{
-			size = currentmapdata.size() - offset;
-		}
-	}
-	memcpy(&currentmapdata[offset], data, size);
-	currentmapdataprocessed = false;
-	//printf("stored map chunk %d %d\n", offset, size);
-}
-
-void World::SendPing(void){
-	Serializer data;
-	Uint8 code = MSG_PING;
-	data.Put(code);
-	lastpingid++;
-	data.Put(lastpingid);
-	SendPacket(GetAuthorityPeer(), data.data, data.BitsToBytes(data.offset));
-	lastpingsent = SDL_GetTicks();
-}
-
-int World::GetPingTime(void){
-	if(state == CONNECTED){
-		return pingtime;
-	}else{
-		return 0;
-	}
-}
-
-int World::AveragePingJitter(void){
-	int average = 0;
-	for(int i = 0; i < sizeof(pinghistory) / sizeof(int); i++){
-		int jitter = abs(pinghistory[i] - pinghistory[(i - 1) % sizeof(pinghistory) / sizeof(int)]);
-		average += jitter;
-	}
-	average /= sizeof(pinghistory) / sizeof(int);
-	return average;
-}
-
-bool World::SecurityIDCanSpawn(Uint8 securityid){
-	// security ids:
-	// 0: all
-	// 1: low
-	// 2: medium
-	// 3: low medium
-	// 4: high
-	// 5: low high
-	// 6: medium high
-	if(gameinfo.securitylevel == LobbyGame::SECNONE){
-		return false;
-	}
-	switch(securityid){
-		case 0:
-			return true;
-		break;
-		case 1:
-			if(gameinfo.securitylevel == LobbyGame::SECLOW){
-				return true;
-			}
-		break;
-		case 2:
-			if(gameinfo.securitylevel == LobbyGame::SECMEDIUM){
-				return true;
-			}
-		break;
-		case 3:
-			if(gameinfo.securitylevel == LobbyGame::SECLOW || gameinfo.securitylevel == LobbyGame::SECMEDIUM){
-				return true;
-			}
-		break;
-		case 4:
-			if(gameinfo.securitylevel == LobbyGame::SECHIGH){
-				return true;
-			}
-		break;
-		case 5:
-			if(gameinfo.securitylevel == LobbyGame::SECLOW || gameinfo.securitylevel == LobbyGame::SECHIGH){
-				return true;
-			}
-		break;
-		case 6:
-			if(gameinfo.securitylevel == LobbyGame::SECMEDIUM || gameinfo.securitylevel == LobbyGame::SECHIGH){
-				return true;
-			}
-		break;
-	}
-	return false;
-}
-
-void World::SetSystemCamera(bool system, Uint16 objectfollow, Sint16 x, Sint16 y){
-	systemcameraactive[system] = true;
-	systemcamerafollow[system] = objectfollow;
-	systemcamerax[system] = x;
-	systemcameray[system] = y;
-}
-
-void World::BroadcastCamera(Sint16 x, Sint16 y){
-	if(!IsAuthority()) return;
-	char msg[5];
-	msg[0] = MSG_CAMERA;
-	memcpy(&msg[1], &x, 2);
-	memcpy(&msg[3], &y, 2);
-	for(unsigned int i = 0; i < maxpeers; i++){
-		Peer * p = peerlist[i];
-		if(p && i != localpeerid) SendPacket(p, msg, 5);
-	}
-}
-
-Object * World::GetObjectFromId(Uint16 id){
-	if(objectidlookup.find(id) != objectidlookup.end()){
-		return objectidlookup[id];
-	}
-	return 0;
-}
-
-void World::SaveSnapshot(Serializer & data, Uint8 peerid){
-	if(mode == AUTHORITY){
-		Player * player = GetPeerPlayer(peerid);
-		bool isobserver = peerlist[peerid] && peerlist[peerid]->observer;
-		Serializer ** oldsnapshotptr = &oldsnapshots[peerid][tickcount % maxoldsnapshots];
-		Serializer ** deltasnapshotptr = &oldsnapshots[peerid][peerlist[peerid]->lasttick % maxoldsnapshots];
-		if(tickcount - peerlist[peerid]->lasttick >= maxoldsnapshots){
-			*deltasnapshotptr = 0;
-		}
-		if(!(*oldsnapshotptr)){
-			*oldsnapshotptr = new Serializer;
-		}
-		
-		// Find deleted objects
-		std::vector<Uint16> deletedobjects;
-		for(int i = peerlist[peerid]->lasttick; i < tickcount; i++){
-			// Go through all the snapshots sent to the peer since the last acknowledged one
-			// and find objects that no longer exist.  Alternative is to create a reliable
-			// packet and just send them when objects are deleted - TODO
-			Serializer ** tempdeltasnapshotptr = &oldsnapshots[peerid][(tickcount - i) % maxoldsnapshots];
-			if(*tempdeltasnapshotptr){
-				(*tempdeltasnapshotptr)->readoffset = 0;
-				while((*tempdeltasnapshotptr)->MoreBytesToRead()){
-					Uint8 type;
-					Uint16 id;
-					int oldreadoffset = (*tempdeltasnapshotptr)->readoffset;
-					(*tempdeltasnapshotptr)->Get(type);
-					(*tempdeltasnapshotptr)->Get(id);
-					(*tempdeltasnapshotptr)->readoffset = oldreadoffset;
-					Object * object = GetObjectFromId(id);
-					if(!object && std::find(deletedobjects.begin(), deletedobjects.end(), id) == deletedobjects.end()){
-						deletedobjects.push_back(id);
-					}
-					(*tempdeltasnapshotptr)->readoffset += objecttypes.SerializedSize(type);
-				}
-			}
-			if(i > maxoldsnapshots){
-				break;
-			}
-		}
-		Uint16 deletedobjectscount = deletedobjects.size();
-		//
-		
-		data.Put(peerlist[peerid]->lasttick);
-		data.Put(deletedobjectscount);
-		for(std::vector<Uint16>::iterator it = deletedobjects.begin(); it != deletedobjects.end(); it++){
-			data.Put(*it);
-		}
-		(*oldsnapshotptr)->offset = 0;
-		
-		if(DELTAENABLED && *deltasnapshotptr){
-			// Write delta'ed objects in snapshot
-			(*deltasnapshotptr)->readoffset = 0;
-			std::map<Uint16, Object *> oldobjects;
-			while((*deltasnapshotptr)->MoreBytesToRead()){
-				Uint8 type;
-				Uint16 id;
-				int oldreadoffset = (*deltasnapshotptr)->readoffset;
-				(*deltasnapshotptr)->Get(type);
-				(*deltasnapshotptr)->Get(id);
-				(*deltasnapshotptr)->readoffset = oldreadoffset;
-				Object * object = GetObjectFromId(id);
-				if(object){
-					oldobjects[id] = object;
-					data.PutBit(1);
-					object->Serialize(Serializer::WRITE, data, *deltasnapshotptr);
-				}else{
-					(*deltasnapshotptr)->readoffset += objecttypes.SerializedSize(type);
-				}
-			}
-			//
-			
-			// Write all new objects
-			for(std::list<Object *>::iterator i = objectlist.begin(); i != objectlist.end(); i++){
-				if((*i)->RequiresAuthority() && (isobserver || RelevantToPlayer(player, (*i)))){
-					(*i)->Serialize(Serializer::WRITE, **oldsnapshotptr);
-					if(oldobjects.find((*i)->id) == oldobjects.end()){
-						data.PutBit(0);
-						(*i)->Serialize(Serializer::WRITE, data);
-					}
-				}
-			}
-			//
-		}else{
-			// Write all relevant objects, no delta compression
-			for(std::list<Object *>::iterator i = objectlist.begin(); i != objectlist.end(); i++){
-				if((*i)->RequiresAuthority() && (isobserver || RelevantToPlayer(player, (*i)))){
-					(*i)->Serialize(Serializer::WRITE, **oldsnapshotptr);
-					data.PutBit(0);
-					(*i)->Serialize(Serializer::WRITE, data);
-				}
-			}
-			//
-		}
-	}else
-	if(mode == REPLICA){
-		Uint32 nulltickcount = 0;
-		data.Put(nulltickcount);
-		Uint16 nulldeletedobjectscount = 0;
-		data.Put(nulldeletedobjectscount);
-		for(std::list<Object *>::iterator i = objectlist.begin(); i != objectlist.end(); i++){
-			if((*i)->RequiresAuthority()){
-				data.PutBit(0);
-				(*i)->Serialize(Serializer::WRITE, data);
-			}
-		}
-	}
-}
-
-void World::LoadSnapshot(Serializer & data, bool create, Serializer * delta, Uint16 objectid){
-	Uint32 deltatick;
-	data.Get(deltatick);
-	Uint16 deletedobjectscount;
-	data.Get(deletedobjectscount);
-	for(int i = 0; i < deletedobjectscount; i++){
-		Uint16 objectid;
-		data.Get(objectid);
-		if(GetObjectFromId(objectid)){
-			//printf("deleted %d\n", objectid);
-			MarkDestroyObject(objectid);
-		}
-	}
-	while(data.MoreBytesToRead()){
-		bool isdeltacompressed = data.GetBit();
-		Uint8 type;
-		Uint16 id;
-		int readoffset = data.readoffset;
-		data.Get(type);
-		data.Get(id);
-		data.readoffset = readoffset;
-		if(objectid && id != objectid){
-			data.readoffset += objecttypes.SerializedSize(type);
-			continue;
-		}
-		Object * object = GetObjectFromId(id);
-		if(!object && create){
-			object = CreateObject(type, id);
-		}
-		if(object){
-			object->lastsnapshottick = tickcount;
-			if(object->type != type){
-				//printf("OBJECT TYPE DOES NOT MATCH IN SNAPSHOT\n");
-				MarkDestroyObject(object->id);
-				data.readoffset += objecttypes.SerializedSize(type);
-			}else{
-				if(object->iscontrollable){
-					//if(!replaying){
-						object->oldx = object->x;
-						object->oldy = object->y;
-					//}
-				}
-				if(delta){
-					delta->readoffset = 0;
-					LoadSnapshot(*delta, false, 0, id);
-				}
-				if(isdeltacompressed){
-					object->Serialize(Serializer::READ, data, (Serializer *)true);
-				}else{
-					object->Serialize(Serializer::READ, data);
-				}
-			}
-		}else{
-			data.readoffset += objecttypes.SerializedSize(type);
-		}
-	}
-}
-
-Object * World::CreateObject(Uint8 type, Uint16 id){
-	if(replaying){ // Do not create objects when rewinding/replaying game state
-		return 0;
-	}
-	if(objectlist.size() == maxobjects){
-		return 0;
-	}
-	Object * object = objecttypes.CreateFromType(type);
-	if(!object){
-		return 0;
-	}
-	if(id == 0 && object->RequiresAuthority() && mode != AUTHORITY){
-		delete object;
-		return 0;
-	}
-	Uint16 searchid = currentid;
-	if(id == 0){
-		if(object->RequiresAuthority()){
-			searchid = currentid;
-		}else{
-			searchid = currentid | 0x8000;
-		}
-		while(currentid <= maxobjects){
-			if(!objectidlookup[searchid]){
-				break;
-			}
-			currentid++;
-			searchid++;
-			if(currentid >= maxobjects){
-				currentid = 1;
-				searchid = 1;
-			}
-		}
-		object->id = searchid;
-	}else{
-		object->id = id;
-	}
-	objectlist.push_back(object);
-	if(IsCollidable(type)){
-		tobjectlist.push_back(object);
-	}
-	objectsbytype[type].push_back(object->id);
-	objectidlookup[object->id] = object;
-	return object;
-}
-
-void World::MarkDestroyObject(Uint16 id){
-	if(replaying){
-		return;
-	}
-	Object * object = GetObjectFromId(id);
-	if(object){
-		object->wasdestroyed = true;
-		objectdestroylist.push_back(id);
-	}
-}
-
-void World::DestroyMarkedObjects(void){
-	for(std::list<Uint16>::iterator i = objectdestroylist.begin(); i != objectdestroylist.end(); i++){
-		DestroyObject((*i));
-	}
-	objectdestroylist.clear();
-}
-
-void World::DestroyObject(Uint16 id){
-	Object * object = GetObjectFromId(id);
-	if(object){
-		objectidlookup.erase(object->id);
-		object->OnDestroy(*this);
-		objectlist.remove(object);
-		if(IsCollidable(object->type)){
-			tobjectlist.remove(object);
-		}
-		std::vector<Uint16>::iterator f = std::find(objectsbytype[object->type].begin(), objectsbytype[object->type].end(), id);
-		if(f != objectsbytype[object->type].end()){
-			objectsbytype[object->type].erase(f);
-		}
-		delete object;
-	}
-}
-
-void World::DestroyAllObjects(void){
-	for(std::list<Object *>::iterator j = objectlist.begin(); j != objectlist.end(); j++){
-		(*j)->OnDestroy(*this);
-		delete (*j);
-	}
-	objectlist.clear();
-	tobjectlist.clear();
-	objectidlookup.clear();
-	objectdestroylist.clear();
-	for(int i = 0; i < ObjectTypes::MAX_OBJECT_TYPE; i++){
-		objectsbytype[i].clear();
-	}
-}
-
-bool World::TestAABB(int x1, int y1, int x2, int y2, Object * object, std::vector<Uint8> & types, bool onlycollidable){
-	int sx1 = 0, sy1 = 0, sx2 = 0, sy2 = 0;
-	object->GetAABB(resources, &sx1, &sy1, &sx2, &sy2);
-	Uint8 type = object->type;
-	if(types.size() > 0 && std::find(types.begin(), types.end(), type) == types.end()){
-		return false;
-	}
-	if(((x1 <= sx1 && x2 >= sx1) || (x1 <= sx2 && x2 >= sx2) || (x1 >= sx1 && x2 <= sx2)) &&
-	   ((y1 <= sy1 && y2 >= sy1) || (y1 <= sy2 && y2 >= sy2) || (y1 >= sy1 && y2 <= sy2))){
-		return true;
-	}
-	return false;
-}
-
-std::vector<Object *> World::TestAABB(int x1, int y1, int x2, int y2, std::vector<Uint8> & types, Uint16 except, Uint16 teamid, bool onlycollidable){
-	std::vector<Object *> objects;
-	for(std::list<Object *>::iterator i = tobjectlist.begin(); i != tobjectlist.end(); i++){
-		Object * object = (*i);
-		if(object->issprite){
-			if(object->id != except){
-				if(!object->isphysical || !onlycollidable || (object->isphysical && object->collidable)){
-					if(!teamid || (teamid && !BelongsToTeam(*object, teamid))){
-						if(TestAABB(x1, y1, x2, y2, object, types)){
-							objects.push_back(object);
-						}
-					}
-				}
-			}
-		}
-	}
-	return objects;
-}
-
-Object * World::TestIncr(int x1, int y1, int x2, int y2, int * xv, int * yv, std::vector<Uint8> & types, Uint16 except, Uint16 teamid){
-	int xb1 = x1 + (*xv < 0 ? *xv : 0);
-	int yb1 = y1 + (*yv < 0 ? *yv : 0);
-	int xb2 = x2 + (*xv > 0 ? *xv : 0);
-	int yb2 = y2 + (*yv > 0 ? *yv : 0);
-	std::vector<Object *> testobjects = TestAABB(xb1, yb1, xb2, yb2, types, except, teamid); // broadphase
-	std::vector<Object *> test;
-	for(std::vector<Object *>::iterator it = testobjects.begin(); it != testobjects.end(); it++){
-		if((*it)->issprite){
-			test.push_back(*it);
-		}
-	}
-	if(test.size() == 0){
-		return 0;
-	}
-	int dx = *xv;
-	int dy = *yv;
-	int step;
-	int error;
-	int oldxv = *xv;
-	int oldyv = *yv;
-	int yv0 = 0;
-	int xv0 = 0;
-	*xv = 0;
-	*yv = 0;
-	float slope = 0;
-	if(dx){
-		slope = (float)dy / dx;
-	}else{
-		if(oldyv > 0){
-			while(*yv < oldyv){
-				for(std::vector<Object *>::iterator i = test.begin(); i != test.end(); i++){
-					Object * object = *i;
-					if(TestAABB(x1 + *xv, y1 + *yv, x2 + *xv, y2 + *yv, object, types)){
-						*yv = yv0;
-						return object;
-					}
-				}
-				yv0 = *yv;
-				(*yv)++;
-			}
-		}else{
-			while(*yv > oldyv){
-				for(std::vector<Object *>::iterator i = test.begin(); i != test.end(); i++){
-					Object * object = *i;
-					if(TestAABB(x1 + *xv, y1 + *yv, x2 + *xv, y2 + *yv, object, types)){
-						*yv = yv0;
-						return object;
-					}
-				}
-				yv0 = *yv;
-				(*yv)--;
-			}
-		}
-	}
-	if(slope > -1 && slope < 1){
-		error = -dx / 2;
-		oldyv > 0 ? step = 1 : step = -1;
-		if(oldxv > 0){
-			while(*xv < oldxv){
-				for(std::vector<Object *>::iterator i = test.begin(); i != test.end(); i++){
-					Object * object = *i;
-					if(TestAABB(x1 + *xv, y1 + *yv, x2 + *xv, y2 + *yv, object, types)){
-						*yv = yv0;
-						*xv = xv0;
-						return object;
-					}
-				}
-				error += dy * step;
-				if(error >= 0){
-					yv0 = *yv;
-					*yv += step;
-					error -= dx;
-				}
-				xv0 = *xv;
-				(*xv)++;
-			}
-		}else{
-			while(*xv > oldxv){
-				for(std::vector<Object *>::iterator i = test.begin(); i != test.end(); i++){
-					Object * object = *i;
-					if(TestAABB(x1 + *xv, y1 + *yv, x2 + *xv, y2 + *yv, object, types)){
-						*yv = yv0;
-						*xv = xv0;
-						return object;
-					}
-				}
-				error += dy * -step;
-				if(error <= 0){
-					yv0 = *yv;
-					*yv += step;
-					error -= dx;
-				}
-				xv0 = *xv;
-				(*xv)--;
-			}
-		}
-	}else{
-		error = -dy / 2;
-		oldxv > 0 ? step = 1 : step = -1;
-		if(oldyv > 0){
-			while(*yv < oldyv){
-				for(std::vector<Object *>::iterator i = test.begin(); i != test.end(); i++){
-					Object * object = *i;
-					if(TestAABB(x1 + *xv, y1 + *yv, x2 + *xv, y2 + *yv, object, types)){
-						*xv = xv0;
-						*yv = yv0;
-						return object;
-					}
-				}
-				error += dx * step;
-				if(error >= 0){
-					xv0 = *xv;
-					*xv += step;
-					error -= dy;
-				}
-				yv0 = *yv;
-				(*yv)++;
-			}
-		}else{
-			while(*yv > oldyv){
-				for(std::vector<Object *>::iterator i = test.begin(); i != test.end(); i++){
-					Object * object = *i;
-					if(TestAABB(x1 + *xv, y1 + *yv, x2 + *xv, y2 + *yv, object, types)){
-						*xv = xv0;
-						*yv = yv0;
-						return object;
-					}
-				}
-				error += dx * -step;
-				if(error <= 0){
-					xv0 = *xv;
-					*xv += step;
-					error -= dy;
-				}
-				yv0 = *yv;
-				(*yv)--;
-			}
-		}
-	}
-	for(std::vector<Object *>::iterator i = test.begin(); i != test.end(); i++){
-		Object * object = *i;
-		if(TestAABB(x1 + *xv, y1 + *yv, x2 + *xv, y2 + *yv, object, types)){
-			*xv = xv0;
-			*yv = yv0;
-			return object;
-		}
-	}
-	return 0;
-}
